[36] feat : 공유옵션 컴포넌트 분리
